refactor(requests): Rename to `ReviewQuoteProposalRequest` class.
Added in QuoteProposals dir.
