+++ app/Http/Controllers/Quotes/Proposals/ReviewQuoteProposalController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\Quotes\Proposals;
 use App\Constants\QuoteProposalTypesConstant;
 use App\Constants\UserTypesAbilitiesConstant;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Quotes\Proposal\CreateQuoteProposalRequest;
+use App\Http\Requests\QuoteProposals\ReviewQuoteProposalRequest;
 use App\Models\QuoteProposal;
 use Illuminate\Http\JsonResponse;
 
@@ -16,7 +16,7 @@ class ReviewQuoteProposalController extends Controller
         $this->middleware(['auth:sanctum', 'abilities:' . UserTypesAbilitiesConstant::ADMIN]);
     }
 
-    public function __invoke(CreateQuoteProposalRequest $request, QuoteProposal $quote_proposal): JsonResponse
+    public function __invoke(ReviewQuoteProposalRequest $request, QuoteProposal $quote_proposal): JsonResponse
     {
         $data = $request->validated();
